split client into few submodules
